requirements.txt
streamlit
openai
google-generativeai==0.3.1
